Split stream encryption into separate EncryptedStreamUtil.
